requirements.txt
pytest-playwright
playwright 